Added Player Controllable Prefab - Removed Example Assets
Added controllable Player Prefab. This includes a greyblock character, orbit and camera. Plug and play if imported correctly. Just drag the prefab which lies in the folder into the scene, and voila!

Removed Example Assets as these are high poly models and drastically reduce the loading efficiency when loading the scene. Hopefully now, it will go alot smoother.